:recycle: refactor: ディレクトリ構成の整理
