+++ src/components/features/sample/CounterDisplay.vue
