+++ src/layouts/default.vue
@@ -1,0 +1,5 @@
+<template>
+  <main>
+    <NuxtPage />
+  </main>
+</template>
