+++ tsconfig.json
@@ -1,12 +1,4 @@
 {
   // https://nuxt.com/docs/guide/concepts/typescript
-  "extends": "./.nuxt/tsconfig.json",
-  "compilerOptions": {
-    "baseUrl": ".",
-    "paths": {
-      "@/*": ["src/*"]
-    },
-    "esModuleInterop": true,
-    "skipLibCheck": true
-  }
+  "extends": "./.nuxt/tsconfig.json"
 }
